Migrate invalid measure scoring test.
