+++ cqf-fhir-cr/src/test/java/org/opencds/cqf/fhir/cr/measure/r4/MeasureAdditionalDataAndCustomOptionsTests.java
@@ -24,7 +24,7 @@ public class MeasureAdditionalDataAndCustomOptionsTests {
         var report = Measure.given()
                 .repositoryFor("AdditionalDataAndCustomOptions")
                 .when()
-                .measureId("GlycemicControlHypoglycemicInitialPopulation")
+                .measureId("SimpleAdditionalDataAndCustomOptions")
                 .periodStart("2022-01-01")
                 .periodEnd("2022-01-31")
                 .subject("Patient/AdditionalDataPatient")
@@ -57,8 +57,8 @@ public class MeasureAdditionalDataAndCustomOptionsTests {
                 .repositoryFor("AdditionalDataAndCustomOptions")
                 .evaluationOptions(evaluationOptions)
                 .when()
-                .measureId("GlycemicControlHypoglycemicInitialPopulation")
-                .subject("Patient/eNeMVHWfNoTsMTbrwWQQ30A3")
+                .measureId("SimpleAdditionalDataAndCustomOptions")
+                .subject("Patient/SimpleAdditionalDataAndCustomOptionsPatient")
                 .periodStart("2022-01-01")
                 .periodEnd("2022-06-29")
                 .reportType("subject")
@@ -91,7 +91,7 @@ public class MeasureAdditionalDataAndCustomOptionsTests {
                 .repositoryFor("AdditionalDataAndCustomOptions")
                 .evaluationOptions(evaluationOptions)
                 .when()
-                .measureId("GlycemicControlHypoglycemicInitialPopulation")
+                .measureId("SimpleAdditionalDataAndCustomOptions")
                 .periodStart("2022-01-01")
                 .periodEnd("2022-01-31")
                 .subject("Patient/AdditionalDataPatient")
