+++ cqf-fhir-cr/src/test/resources/org/opencds/cqf/fhir/cr/measure/r4/AdditionalDataAndCustomOptions/cql/SimpleAdditionalDataAndCustomOptions.cql
@@ -1,4 +1,4 @@
-library GlycemicControlHypoglycemicInitialPopulation version '0.0.004'
+library SimpleAdditionalDataAndCustomOptions version '0.0.004'
 
 using FHIR version '4.0.1'
