+++ cqf-fhir-cr/src/test/resources/org/opencds/cqf/fhir/cr/measure/r4/AdditionalDataAndCustomOptions/resources/library/SimpleAdditionalDataAndCustomOptions.json
@@ -1,6 +1,6 @@
 {
   "resourceType": "Library",
-  "id": "GlycemicControlHypoglycemicInitialPopulation",
+  "id": "SimpleAdditionalDataAndCustomOptions",
   "meta": {
     "profile": [ "http://hl7.org/fhir/us/cqfmeasures/StructureDefinition/library-cqfm" ]
   },
@@ -8,9 +8,9 @@
     "url": "http://hl7.org/fhir/us/cqfmeasures/StructureDefinition/cqfm-populationBasis",
     "valueCode": "Encounter"
   } ],
-  "url": "http://test-case.com/fhir/measures/Library/GlycemicControlHypoglycemicInitialPopulation",
+  "url": "http://test-case.com/fhir/measures/Library/SimpleAdditionalDataAndCustomOptions",
   "version": "0.0.004",
-  "name": "GlycemicControlHypoglycemicInitialPopulation",
+  "name": "SimpleAdditionalDataAndCustomOptions",
   "status": "draft",
   "experimental": false,
   "type": {
@@ -20,9 +20,8 @@
       "display": "Logic Library"
     } ]
   },
-  "description": "All inpatient encounters (including ED/Observation visits that end within 1 hour of the start of the inpatient encounter) for patients of all ages where at least one ADD was ordered or administered during the encounter that is during the measurement period.",
   "content": [ {
     "contentType": "text/cql",
-    "url": "../../cql/GlycemicControlHypoglycemicInitialPopulation.cql"
+    "url": "../../cql/SimpleAdditionalDataAndCustomOptions.cql"
   } ]
 }
