--- cqf-fhir-cr/src/test/resources/org/opencds/cqf/fhir/cr/measure/r4/AdditionalDataAndCustomOptions/valueset/2.16.840.1.113762.1.4.1190.38.json
@@ -1,169 +1,0 @@
-{
-  "resourceType": "ValueSet",
-  "id": "2.16.840.1.113762.1.4.1190.38",
-  "meta": {
-    "versionId": "18",
-    "lastUpdated": "2022-02-10T01:02:28.000-05:00"
-  },
-  "url": "http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113762.1.4.1190.38",
-  "version": "20220210",
-  "name": "Blood Glucose Laboratory and Point of Care Tests",
-  "title": "Blood Glucose Laboratory and Point of Care Tests",
-  "status": "active",
-  "date": "2022-02-10T01:02:28-05:00",
-  "publisher": "HL7 Terminology",
-  "expansion": {
-    "identifier": "urn:uuid:62c58d8b-7617-482a-96cd-e84fb703d2ed",
-    "timestamp": "2022-03-09T10:16:01-05:00",
-    "total": 28,
-    "offset": 0,
-    "parameter": [ {
-      "name": "count",
-      "valueInteger": 1000
-    }, {
-      "name": "offset",
-      "valueInteger": 0
-    } ],
-    "contains": [ {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "10449-7",
-      "display": "Glucose [Mass/volume] in Serum or Plasma --1 hour post meal"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "10450-5",
-      "display": "Glucose [Mass/volume] in Serum or Plasma --10 hours fasting"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "1521-4",
-      "display": "Glucose [Mass/volume] in Serum or Plasma --2 hours post meal"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "1547-9",
-      "display": "Glucose [Mass/volume] in Serum or Plasma --baseline"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "1550-3",
-      "display": "Glucose [Mass/volume] in Serum or Plasma --pre 12 hour fast"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "1554-5",
-      "display": "Glucose [Mass/volume] in Serum or Plasma --12 hours fasting"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "1556-0",
-      "display": "Fasting glucose [Mass/volume] in Capillary blood"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "1557-8",
-      "display": "Fasting glucose [Mass/volume] in Venous blood"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "1558-6",
-      "display": "Fasting glucose [Mass/volume] in Serum or Plasma"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "16915-1",
-      "display": "Glucose [Mass/volume] in Serum or Plasma --post meal"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "17865-7",
-      "display": "Glucose [Mass/volume] in Serum or Plasma --8 hours fasting"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "2339-0",
-      "display": "Glucose [Mass/volume] in Blood"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "2340-8",
-      "display": "Glucose [Mass/volume] in Blood by Automated test strip"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "2341-6",
-      "display": "Glucose [Mass/volume] in Blood by Test strip manual"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "2345-7",
-      "display": "Glucose [Mass/volume] in Serum or Plasma"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "32016-8",
-      "display": "Glucose [Mass/volume] in Capillary blood"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "40858-3",
-      "display": "Glucose [Mass/volume] in Capillary blood --baseline"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "41604-0",
-      "display": "Fasting glucose [Mass/volume] in Capillary blood by Glucometer"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "41651-1",
-      "display": "Glucose [Mass/volume] in Arterial blood"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "41652-9",
-      "display": "Glucose [Mass/volume] in Venous blood"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "41653-7",
-      "display": "Glucose [Mass/volume] in Capillary blood by Glucometer"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "53049-3",
-      "display": "Glucose [Mass/volume] in Serum or Plasma --pre-meal"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "6689-4",
-      "display": "Glucose [Mass/volume] in Blood --2 hours post meal"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "74774-1",
-      "display": "Glucose [Mass/volume] in Serum, Plasma or Blood"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "80959-0",
-      "display": "Glucose [Mass/volume] in Serum or Plasma --1.5 hours post meal"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "87422-2",
-      "display": "Glucose [Mass/volume] in Blood --post meal"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "88365-2",
-      "display": "Glucose [Mass/volume] in Blood --pre-meal"
-    }, {
-      "system": "http://loinc.org",
-      "version": "2.71",
-      "code": "93791-2",
-      "display": "Glucose [Mass/volume] mean in Serum or Plasma"
-    } ]
-  }
-}
