--- cqf-fhir-cr/src/test/resources/org/opencds/cqf/fhir/cr/measure/r4/AdditionalDataAndCustomOptions/valueset/2.16.840.1.113883.3.117.1.7.1.292.json
@@ -1,35 +1,0 @@
-{
-  "resourceType": "ValueSet",
-  "id": "2.16.840.1.113883.3.117.1.7.1.292",
-  "meta": {
-    "versionId": "34",
-    "lastUpdated": "2021-06-11T01:02:23.000-04:00"
-  },
-  "url": "http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.113883.3.117.1.7.1.292",
-  "version": "20210611",
-  "name": "Emergency Department Visit",
-  "title": "Emergency Department Visit",
-  "status": "active",
-  "date": "2021-06-11T01:02:23-04:00",
-  "publisher": "TJC EH Steward",
-  "description": "The purpose of this value set is to represent concepts for encounters in the emergency department (ED).",
-  "expansion": {
-    "identifier": "urn:uuid:8a4ca8c5-3324-49c6-b9f3-61b45192e630",
-    "timestamp": "2022-03-09T10:10:51-05:00",
-    "total": 1,
-    "offset": 0,
-    "parameter": [ {
-      "name": "count",
-      "valueInteger": 1000
-    }, {
-      "name": "offset",
-      "valueInteger": 0
-    } ],
-    "contains": [ {
-      "system": "http://snomed.info/sct",
-      "version": "http://snomed.info/sct/731000124108/version/20210901",
-      "code": "4525004",
-      "display": "Emergency department patient visit (procedure)"
-    } ]
-  }
-}
