--- cqf-fhir-cr/src/test/resources/org/opencds/cqf/fhir/cr/measure/r4/AdditionalDataAndCustomOptions/valueset/2.16.840.1.114222.4.11.836.json
@@ -1,60 +1,0 @@
-{
-  "resourceType": "ValueSet",
-  "id": "2.16.840.1.114222.4.11.836",
-  "meta": {
-    "versionId": "1",
-    "lastUpdated": "2012-10-25T12:28:31.000-04:00"
-  },
-  "url": "http://cts.nlm.nih.gov/fhir/ValueSet/2.16.840.1.114222.4.11.836",
-  "version": "20121025",
-  "name": "Race",
-  "title": "Race",
-  "status": "active",
-  "date": "2012-10-25T12:28:31-04:00",
-  "publisher": "CDC NCHS",
-  "description": "The purpose of this value set is to represent CDC concepts for Race.",
-  "expansion": {
-    "identifier": "urn:uuid:55002dc9-9101-423c-9850-a8b376412366",
-    "timestamp": "2022-03-15T14:18:47-04:00",
-    "total": 6,
-    "offset": 0,
-    "parameter": [ {
-      "name": "count",
-      "valueInteger": 1000
-    }, {
-      "name": "offset",
-      "valueInteger": 0
-    } ],
-    "contains": [ {
-      "system": "urn:oid:2.16.840.1.113883.6.238",
-      "version": "1.2",
-      "code": "1002-5",
-      "display": "American Indian or Alaska Native"
-    }, {
-      "system": "urn:oid:2.16.840.1.113883.6.238",
-      "version": "1.2",
-      "code": "2028-9",
-      "display": "Asian"
-    }, {
-      "system": "urn:oid:2.16.840.1.113883.6.238",
-      "version": "1.2",
-      "code": "2054-5",
-      "display": "Black or African American"
-    }, {
-      "system": "urn:oid:2.16.840.1.113883.6.238",
-      "version": "1.2",
-      "code": "2076-8",
-      "display": "Native Hawaiian or Other Pacific Islander"
-    }, {
-      "system": "urn:oid:2.16.840.1.113883.6.238",
-      "version": "1.2",
-      "code": "2106-3",
-      "display": "White"
-    }, {
-      "system": "urn:oid:2.16.840.1.113883.6.238",
-      "version": "1.2",
-      "code": "2131-1",
-      "display": "Other Race"
-    } ]
-  }
-}
