+++ cqf-fhir-cr/src/test/resources/org/opencds/cqf/fhir/cr/measure/r4/MeasureScoring/cql/MeasureScoringInvalid.cql
@@ -1,0 +1,5 @@
+library MeasureScoringInvalid version '1.0.001'
+
+using FHIR version '4.0.1'
+
+include FHIRHelpers version '4.0.001' called FHIRHelpers
