+++ cqf-fhir-cr/src/test/resources/org/opencds/cqf/fhir/cr/measure/r4/MeasureScoring/resources/library/DischargedonAntithromboticTherapyFHIR-Valid.json
@@ -1,0 +1,29 @@
+{
+  "resourceType": "Library",
+  "id": "DischargedonAntithromboticTherapyFHIR-Valid",
+  "meta": {
+    "versionId": "1",
+    "lastUpdated": "2021-05-17T10:33:50.000-06:00",
+    "source": "#S5FYKrTuRBHdxSm2",
+    "profile": [ "http://hl7.org/fhir/us/cqfmeasures/StructureDefinition/computable-library-cqfm" ]
+  },
+  "language": "en",
+  "url": "https://madie.cms.gov/Library/DischargedonAntithromboticTherapyFHIR-Valid",
+  "version": "1.0.001",
+  "name": "DischargedonAntithromboticTherapyFHIR-Valid",
+  "status": "active",
+  "experimental": false,
+  "type": {
+    "coding": [ {
+      "system": "http://terminology.hl7.org/CodeSystem/library-type",
+      "code": "logic-library"
+    } ]
+  },
+  "date": "2021-05-17T16:33:49+00:00",
+  "publisher": "The Joint Commission",
+  "description": "Discharged on Antithrombotic Therapy FHIR",
+  "content": [ {
+    "contentType": "text/cql",
+    "url": "../../cql/DischargedonAntithromboticTherapyFHIR-Valid.cql"
+  } ]
+}
