+++ cqf-fhir-cr/src/test/resources/org/opencds/cqf/fhir/cr/measure/r4/MeasureScoring/tests/encounter/numer-EXM104-2.json
@@ -1,0 +1,40 @@
+{
+  "resourceType": "Encounter",
+  "id": "numer-EXM104-2",
+  "meta": {
+    "profile": [ "http://hl7.org/fhir/us/core/StructureDefinition/us-core-encounter" ]
+  },
+  "status": "finished",
+  "class": {
+    "system": "http://terminology.hl7.org/CodeSystem/v3-ActCode",
+    "code": "IMP",
+    "display": "inpatient encounter"
+  },
+  "type": [ {
+    "coding": [ {
+      "system": "http://snomed.info/sct",
+      "code": "32485007",
+      "display": "Hospital admission (procedure)"
+    } ]
+  } ],
+  "subject": {
+    "reference": "Patient/numer-EXM104"
+  },
+  "period": {
+    "start": "2019-08-21T00:00:00-06:00",
+    "end": "2019-12-19T08:15:00-07:00"
+  },
+  "diagnosis": [ {
+    "condition": {
+      "reference": "Condition/numer-EXM104-1"
+    },
+    "use": {
+      "coding": [ {
+        "system": "http://hl7.org/fhir/diagnosis-role",
+        "code": "billing",
+        "display": "Billing"
+      } ]
+    },
+    "rank": 1
+  } ]
+}
