+++ cqf-fhir-cr/src/test/resources/org/opencds/cqf/fhir/cr/measure/r4/MeasureScoring/tests/medicationrequest/numer-EXM104-5.json
@@ -1,0 +1,24 @@
+{
+  "resourceType": "MedicationRequest",
+  "id": "numer-EXM104-5",
+  "status": "completed",
+  "intent": "order",
+  "category": [ {
+    "coding": [ {
+      "system": "http://terminology.hl7.org/CodeSystem/medicationrequest-category",
+      "code": "discharge",
+      "display": "Discharge"
+    } ]
+  } ],
+  "medicationCodeableConcept": {
+    "coding": [ {
+      "system": "http://www.nlm.nih.gov/research/umls/rxnorm",
+      "code": "1037045",
+      "display": "dabigatran etexilate 150 MG Oral Capsule"
+    } ]
+  },
+  "subject": {
+    "reference": "Patient/numer-EXM104"
+  },
+  "authoredOn": "2019-12-17T08:00:00"
+}
